instant search

+++ app/Http/Controllers/API/UserController.php
@@ -73,8 +73,13 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+       // $this->authorize('isAdmin');
+        if(\Gate::allows('isAdmin') || \Gate::allows('isAuthor')){
+            return User::latest()->paginate(4);
+        }
+        
     }
+
 
     /**
      * Store a newly created resource in storage.
@@ -139,10 +144,31 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
+
         $user = User::findOrFail($id);
 
         $user->delete();
 
         return ['message' => 'User deleted'];
     }
+
+
+
+    public function search(){
+
+        if($search = \Request::get('q')){
+            $users = User::where(function($query) use ($search) {
+                $query->where('name','LIKE',"%$search%")
+                        ->orWhere('email','LIKE',"%$search%")
+                        ->orWhere('type','LIKE',"%$search%");
+            })->paginate(10);
+        }else{
+            $users = User::latest()->paginate(4);
+        }
+        return $users;
+    }
+
+
+
 }
